requirement.txt
fastapi
uvicorn
python-multipart
pdf2image
pillow
pytesseract